Revert "Modal til navn"
