+++ package.json
@@ -21,7 +21,7 @@
     "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.11.0",
-    "react-router-dom": "^6.17.0",
+    "react-router-dom": "^6.16.0",
     "react-scripts": "^5.0.1",
     "save": "^2.9.0"
   },
