+++ src/pages/StatisticsPage.jsx
@@ -10,6 +10,8 @@ export default function StatisticsPage(){
 
         if (temp) {
             setBrugernavn(temp)
+        } else {
+            setBrugernavn("you");
         }
 
     },[])
